Tea-8707: Dette med å flytte resultata er ei ganske atskilt problemstilling, så synes det høyrer heime i ei eiga fil. Utils-objekt/-klassar har generelt ein lei tendens til å vekse og vekse, og ha ansvar for ein gasillion forskjellige ting. Synes det er ryddigare med separate objekt
