+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/vilkårsvurdering/VilkårsvurderingUtils.kt
@@ -28,7 +28,6 @@ import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.Vilkår
 import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.VilkårResultat
 import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.Vilkårsvurdering
 import java.time.LocalDate
-import java.time.Period
 
 object VilkårsvurderingUtils {
 
@@ -199,122 +198,13 @@ object VilkårsvurderingUtils {
         aktivVilkårsvurdering: Vilkårsvurdering,
         forrigeBehandlingVilkårsvurdering: Vilkårsvurdering? = null,
         løpendeUnderkategori: BehandlingUnderkategori? = null
-    ): Pair<Vilkårsvurdering, Vilkårsvurdering> {
-
-        // OBS!! MÅ jobbe på kopier av vilkårsvurderingen her for å ikke oppdatere databasen
-        // Viktig at det er vår egen implementasjon av kopier som brukes, da kotlin sin copy-funksjon er en shallow copy
-        val initiellVilkårsvurderingKopi = initiellVilkårsvurdering.kopier()
-        val aktivVilkårsvurderingKopi = aktivVilkårsvurdering.kopier()
-
-        // Identifiserer hvilke vilkår som skal legges til og hvilke som kan fjernes
-        val personResultaterAktivt = aktivVilkårsvurderingKopi.personResultater.toMutableSet()
-        val personResultaterOppdatert = mutableSetOf<PersonResultat>()
-        initiellVilkårsvurderingKopi.personResultater.forEach { personFraInit ->
-            val personTilOppdatert = PersonResultat(
-                vilkårsvurdering = initiellVilkårsvurderingKopi,
-                aktør = personFraInit.aktør
-            )
-            val personenSomFinnes = personResultaterAktivt.firstOrNull { it.aktør == personFraInit.aktør }
-
-            if (personenSomFinnes == null) {
-                // Legg til ny person
-                personTilOppdatert.setSortedVilkårResultater(
-                    personFraInit.vilkårResultater.map { it.kopierMedParent(personTilOppdatert) }
-                        .toSet()
-                )
-            } else {
-                // Fyll inn den initierte med person fra aktiv
-                oppdaterEksisterendePerson(
-                    personenSomFinnes = personenSomFinnes,
-                    personFraInit = personFraInit,
-                    kopieringSkjerFraForrigeBehandling = initiellVilkårsvurderingKopi.behandling.id != aktivVilkårsvurderingKopi.behandling.id,
-                    personTilOppdatert = personTilOppdatert,
-                    forrigeBehandlingVilkårsvurdering = forrigeBehandlingVilkårsvurdering,
-                    løpendeUnderkategori = løpendeUnderkategori,
-                    personResultaterAktivt = personResultaterAktivt
-                )
-            }
-            personResultaterOppdatert.add(personTilOppdatert)
-        }
-
-        aktivVilkårsvurderingKopi.personResultater = personResultaterAktivt
-        initiellVilkårsvurderingKopi.personResultater = personResultaterOppdatert
-
-        return Pair(initiellVilkårsvurderingKopi, aktivVilkårsvurderingKopi)
-    }
-
-    private fun oppdaterEksisterendePerson(
-        personenSomFinnes: PersonResultat,
-        personFraInit: PersonResultat,
-        kopieringSkjerFraForrigeBehandling: Boolean,
-        personTilOppdatert: PersonResultat,
-        forrigeBehandlingVilkårsvurdering: Vilkårsvurdering?,
-        løpendeUnderkategori: BehandlingUnderkategori?,
-        personResultaterAktivt: MutableSet<PersonResultat>
-    ) {
-
-        val personsVilkårAktivt = personenSomFinnes.vilkårResultater.toMutableSet()
-        val personsVilkårOppdatert = mutableSetOf<VilkårResultat>()
-        personFraInit.vilkårResultater.forEach { vilkårFraInit ->
-            val vilkårSomFinnes =
-                personenSomFinnes.vilkårResultater.filter { it.vilkårType == vilkårFraInit.vilkårType }
-
-            val vilkårSomSkalKopieresOver = vilkårSomFinnes.filtrerVilkårÅKopiere(
-                kopieringSkjerFraForrigeBehandling = kopieringSkjerFraForrigeBehandling
-            )
-            val vilkårSomSkalFjernesFraAktivt = vilkårSomFinnes - vilkårSomSkalKopieresOver
-            personsVilkårAktivt.removeAll(vilkårSomSkalFjernesFraAktivt)
-
-            if (vilkårSomSkalKopieresOver.isEmpty()) {
-                // Legg til nytt vilkår på person
-                personsVilkårOppdatert.add(vilkårFraInit.kopierMedParent(personTilOppdatert))
-            } else {
-                /*  Vilkår er vurdert på person - flytt fra aktivt og overskriv initierte
-                            ikke oppfylte eller ikke vurdert perioder skal ikke kopieres om minst en oppfylt
-                            periode eksisterer. */
-
-                personsVilkårOppdatert.addAll(
-                    vilkårSomSkalKopieresOver.map { it.kopierMedParent(personTilOppdatert) }
-                )
-                personsVilkårAktivt.removeAll(vilkårSomSkalKopieresOver)
-            }
-        }
-        val eksistererUtvidetVilkårPåForrigeBehandling =
-            forrigeBehandlingVilkårsvurdering?.personResultater
-                ?.firstOrNull { it.aktør == personFraInit.aktør }
-                ?.vilkårResultater
-                ?.any {
-                    it.vilkårType == Vilkår.UTVIDET_BARNETRYGD &&
-                        it.resultat == Resultat.OPPFYLT &&
-                        // forrige behandling har minst et måned ubetalt utvidet barnetrygd
-                        it.differanseIPeriode().toTotalMonths() >= Period.ofMonths(1).months
-                } ?: false
-
-        // Hvis forrige behandling inneholdt utvidet-vilkåret eller underkategorien er utvidet skal
-        // utvidet-vilkåret kopieres med videre uansett nåværende underkategori
-        if (personsVilkårOppdatert.none { vilkårResultat -> vilkårResultat.vilkårType == Vilkår.UTVIDET_BARNETRYGD } &&
-            (eksistererUtvidetVilkårPåForrigeBehandling || løpendeUnderkategori == BehandlingUnderkategori.UTVIDET)
-        ) {
-            val utvidetVilkår =
-                personenSomFinnes.vilkårResultater.filter { vilkårResultat -> vilkårResultat.vilkårType == Vilkår.UTVIDET_BARNETRYGD }
-            if (utvidetVilkår.isNotEmpty()) {
-                personsVilkårOppdatert.addAll(
-                    utvidetVilkår.filtrerVilkårÅKopiere(kopieringSkjerFraForrigeBehandling = kopieringSkjerFraForrigeBehandling)
-                        .map { it.kopierMedParent(personTilOppdatert) }
-                )
-                personsVilkårAktivt.removeAll(utvidetVilkår)
-            }
-        }
-
-        personTilOppdatert.setSortedVilkårResultater(personsVilkårOppdatert.toSet())
-
-        // Fjern person fra aktivt dersom alle vilkår er fjernet, ellers oppdater
-        if (personsVilkårAktivt.isEmpty()) {
-            personResultaterAktivt.remove(personenSomFinnes)
-        } else {
-            personenSomFinnes.setSortedVilkårResultater(personsVilkårAktivt.toSet())
-        }
-    }
+    ): Pair<Vilkårsvurdering, Vilkårsvurdering> =
+        VilkårsvurderingResultatFlytter.flyttResultaterTilInitielt(
+            initiellVilkårsvurdering = initiellVilkårsvurdering,
+            aktivVilkårsvurdering = aktivVilkårsvurdering,
+            forrigeBehandlingVilkårsvurdering = forrigeBehandlingVilkårsvurdering,
+            løpendeUnderkategori = løpendeUnderkategori
+        )
 
     fun lagFjernAdvarsel(personResultater: Set<PersonResultat>): String {
         var advarsel =
@@ -360,7 +250,7 @@ fun eøsStandardbegrunnelserTilNedtrekksmenytekster(
 
 fun vedtakBegrunnelseTilRestVedtakBegrunnelseTilknyttetVilkår(
     sanityBegrunnelser: List<SanityBegrunnelse>,
-    vedtakBegrunnelse: Standardbegrunnelse,
+    vedtakBegrunnelse: Standardbegrunnelse
 ): List<RestVedtakBegrunnelseTilknyttetVilkår> {
     val sanityBegrunnelse = vedtakBegrunnelse.tilSanityBegrunnelse(sanityBegrunnelser) ?: return emptyList()
 
@@ -399,14 +289,6 @@ fun eøsBegrunnelseTilRestVedtakBegrunnelseTilknyttetVilkår(
             vilkår = null
         )
     )
-}
-
-private fun List<VilkårResultat>.filtrerVilkårÅKopiere(kopieringSkjerFraForrigeBehandling: Boolean): List<VilkårResultat> {
-    return if (kopieringSkjerFraForrigeBehandling) {
-        this.filter { it.resultat == Resultat.OPPFYLT }
-    } else {
-        this
-    }
 }
 
 fun genererPersonResultatForPerson(
@@ -463,8 +345,9 @@ private fun utledResultat(
 }
 
 private fun utledResultatForGiftPartnerskap(person: Person) =
-    if (person.sivilstander.isEmpty() || person.sivilstander.sisteSivilstand()?.type?.somForventetHosBarn() == true)
-        Resultat.OPPFYLT else Resultat.IKKE_VURDERT
+    if (person.sivilstander.isEmpty() || person.sivilstander.sisteSivilstand()?.type?.somForventetHosBarn() == true) {
+        Resultat.OPPFYLT
+    }else Resultat.IKKE_VURDERT
 
 private fun utledBegrunnelse(
     vilkår: Vilkår,
@@ -472,9 +355,10 @@ private fun utledBegrunnelse(
 ) = when {
     person.erDød() -> "Dødsfall"
     vilkår == Vilkår.UNDER_18_ÅR -> "Vurdert og satt automatisk"
-    vilkår == Vilkår.GIFT_PARTNERSKAP -> if (person.sivilstander.sisteSivilstand()?.type?.somForventetHosBarn() == false)
+    vilkår == Vilkår.GIFT_PARTNERSKAP -> if (person.sivilstander.sisteSivilstand()?.type?.somForventetHosBarn() == false) {
         "Vilkåret er forsøkt behandlet automatisk, men barnet er registrert som gift i " +
-            "folkeregisteret. Vurder hvilke konsekvenser dette skal ha for behandlingen" else ""
+            "folkeregisteret. Vurder hvilke konsekvenser dette skal ha for behandlingen"
+    }else ""
 
     else -> ""
 }
